refactor: use `rlang::on_load()` to generate glycan graphs for known motifs
In commit ee937f0f, a glycan graph column is added to the `glygen_motifs` data. This has two problems. Firstly, if anything changes in `glyrepr` or `glyparse` packages, the glycan graphs in the rda file may be stale. Using the "on_load" method, the graph objects will always be up-to-date, as long as the underlying IUPAC-condensed structure strings do not change. Secondly, due to the lazy load mechanism of R's internal data management, `glygen_motifs` are sometimes `data.frame` instead of `tibble`, and the "graph" column contains only raw lists instead of "glycan_graph" S3 class objects. The "on_load" method makes sure that everything is in the right types.