perf: speed up algorithms 1000x
Finally, we pinpoint the code causing the performance problem. We used `glyrepr::convert_mono_type()` to ensure that mono types of glycans and motifs are the same. That function uses `glyrepr::spmap_structure()` under the hood, which calls time-consuming `glyrepr:::.structure_to_iupac_single()` to ensure the new `glyrepr::glycan_structure()` has the correct IUPACs. The last step is really time-consuming, but in our case we don't need the new IUPACs at all!

To resolve this, we implement `fast_convert_mono_type`(), a faster version of `glyrepr::convert_mono_type()` for this special use.