feat: redesign quantify_motifs
Redesign `quantify_motifs()` to make it more like `glydet::derive_traits()`. The previous implementation of `quantify_motifs()` had a serious bug: when description columns about glycan structures exist in the variable information tibble, the aggregation logic will fail, and weird results will be generated. The new implementation explicitly relies on the unique combination of `protein` and `protein_site` to form a "glycosite", and calculates motif quantifications on each glycosite. Then, it mimics `quantify_motifs()` to add back columns in the old variable information tibble with "many-to-one" relationship with glycosites. The documentation has also been revised to make it more consistent with `derive_traits()`.